Add unit test, more cleanup
